+++ bs-default-config.php
@@ -1,0 +1,7 @@
+<?php
+
+/* Site language */
+define( 'BS_LANG', 'fr_FR' );
+
+/* Site Name */
+define( 'BS_NAME', 'BackStarter Demo' );