+++ inc/bs-config.php
@@ -9,7 +9,7 @@ $constants = array(
     'BS_LANG' => 'en_EN',
     /* Default Site Name */
     'BS_NAME' => 'BackStarter',
-    /* MySQL */
+    /* Database */
     'BS_DBHOST' => '',
     'BS_DBNAME' => '',
     'BS_DBUSER' => '',