+++ inc/bs-control.php
@@ -5,22 +5,26 @@ if ( !defined( 'BS_PATH' ) ) {
 }
 
 /* ----------------------------------------------------------
+  Models
+---------------------------------------------------------- */
+
+include BS_INC_DIR . 'classes/bs-database.php';
+include BS_INC_DIR . 'classes/bs-user.php';
+include BS_INC_DIR . 'classes/bs-page.php';
+
+/* ----------------------------------------------------------
   Config
 ---------------------------------------------------------- */
 
 if ( file_exists( BS_PATH . 'bs-config.php' ) ) {
     include BS_PATH . 'bs-config.php';
 }
+else {
+    include BS_INC_DIR . 'bs-install.php';
+    die;
+}
 
 include BS_INC_DIR . 'bs-config.php';
-
-/* ----------------------------------------------------------
-  Models
----------------------------------------------------------- */
-
-include BS_INC_DIR . 'classes/bs-mysql.php';
-include BS_INC_DIR . 'classes/bs-user.php';
-include BS_INC_DIR . 'classes/bs-page.php';
 
 /* ----------------------------------------------------------
   Launch project