+++ changelog.md
@@ -30,3 +30,10 @@ Changelog
 * DB Prefix
 * Better MVC organisation
 * Better Model
+
+2014-01-08 : v 0.4.1
+---
+
+* bs-config generation
+* database connection
+* installation start